app/requirements.txt
alembic==1.16.5
asn1crypto==1.5.1
blinker==1.9.0
cbor2==5.7.0
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==25.1.0
python-dotenv==1.1.1
redis==6.4.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
waitress==3.0.0
webauthn==2.7.0
Werkzeug==3.1.3
gunicorn==22.0.0